refactor(web): move lobby buttons into their own components
